Wrap up CryptoOpenSSL_C
- Repurpose zd_truncate to more generic zd_resize
- Add CryptoErrorNone (0)
- Rename _ctx_t suffix to _ctx
- Rename _t enums to _code
- Rearrange test extensions better
- Clean up schemes
