+++ Tests/OpenVPN/CryptoOpenSSL/Legacy/Extensions+Legacy.swift
@@ -76,12 +76,3 @@ extension Optional where Wrapped == CryptoFlagsWrapper {
         }
     }
 }
-
-extension UnsafeRawBufferPointer {
-    var bytePointer: UnsafePointer<Element> {
-        guard let address = bindMemory(to: Element.self).baseAddress else {
-            fatalError("Cannot bind to self")
-        }
-        return address
-    }
-}
