+++ Tests/OpenVPN/CryptoOpenSSL/Native/Extensions+Native.swift
@@ -54,45 +54,3 @@ extension Optional where Wrapped == CryptoFlagsWrapper {
         }
     }
 }
-
-extension Encrypter {
-    func encryptData(_ data: Data, flags: CryptoFlagsWrapper?) throws -> Data {
-        let srcLength = data.count
-        var dest: [UInt8] = Array(repeating: 0, count: srcLength + 256)
-        var destLength = 0
-        _ = try data.withUnsafeBytes {
-            try encryptBytes($0.bytePointer, length: srcLength, dest: &dest, destLength: &destLength, flags: flags)
-        }
-        dest.removeSubrange(destLength..<dest.count)
-        return Data(dest)
-    }
-}
-
-extension Decrypter {
-    func decryptData(_ data: Data, flags: CryptoFlagsWrapper?) throws -> Data {
-        let srcLength = data.count
-        var dest: [UInt8] = Array(repeating: 0, count: srcLength + 256)
-        var destLength = 0
-        _ = try data.withUnsafeBytes {
-            try decryptBytes($0.bytePointer, length: srcLength, dest: &dest, destLength: &destLength, flags: flags)
-        }
-        dest.removeSubrange(destLength..<dest.count)
-        return Data(dest)
-    }
-
-    func verifyData(_ data: Data, flags: CryptoFlagsWrapper?) throws {
-        let srcLength = data.count
-        _ = try data.withUnsafeBytes {
-            try verifyBytes($0.bytePointer, length: srcLength, flags: flags)
-        }
-    }
-}
-
-extension UnsafeRawBufferPointer {
-    var bytePointer: UnsafePointer<Element> {
-        guard let address = bindMemory(to: Element.self).baseAddress else {
-            fatalError("Cannot bind to self")
-        }
-        return address
-    }
-}
